refactor(ssml): move integration scenarios to rust/tests/ssml_integration.rs (#267 F8)
Splits the production-vs-tests boundary in `rust/src/tts/ssml/mod.rs` —
the file was 70% test code (556 LOC of tests vs 315 LOC of production).
Lifts the 34 full-`<speak>…</speak>` parse-and-assert scenarios into a
new external integration test, keeping only the `tag_name()` regression
(needs `pub(super)` access) colocated.

Mechanics:

- rust/tests/ssml_integration.rs (new, 415 LOC) — every test calls
  `parse()` end-to-end and asserts `Vec<Segment>` shape. Imports
  `kesha_engine::tts::ssml::{parse, Segment}` via lib.rs's `pub mod tts`
  re-export, gated `#![cfg(feature = "tts")]`.

- rust/src/tts/ssml/mod.rs — drops the relocated tests, leaves a
  one-test in-crate block with a pointer comment. `DEFAULT_BREAK` stays
  `pub(super)`; the integration test hardcodes 250ms with a comment
  treating the default-break duration as part of the contract — if the
  upstream constant changes, the test alarms (intentional behaviour).

Verification:

- `cargo nextest run --features tts --lib ssml::` → 11 in-crate unit
  tests pass (8 rate + 2 segment + 1 tag_name).
- `cargo nextest run --features tts --test ssml_integration` → all 34
  moved tests pass, 0 skipped.
- `cargo nextest run --features tts` → 415 tests pass, 0 fail
  (includes existing tts_e2e, tts_ru_normalize, prosody_rate, etc.).
- `cargo clippy --all-targets --features tts -- -D warnings` clean.
- `cargo fmt --check` clean.

No behaviour change: pure test re-organisation. SSML parser surface,
warn-once mechanics, segment shapes, error messages all identical.

Refs #267 (F8 follow-up — completes the part that #272 deferred when
it colocated the unit tests).
